+++ fakeredis/commands_mixins/connection_mixin.py
@@ -1,4 +1,4 @@
-from typing import Any
+from typing import Any, List, Union
 
 from fakeredis import _msgs as msgs
 from fakeredis._commands import command, DbIndex
@@ -8,20 +8,22 @@ PONG = SimpleString(b"PONG")
 
 
 class ConnectionCommandsMixin:
-    _server: Any
-    _db: Database
-    _db_num: int
-    _pubsub: int
-
     # Connection commands
     # TODO: auth, quit
 
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
+        super(ConnectionCommandsMixin, self).__init__(*args, **kwargs)
+        self._db: Database
+        self._db_num: int
+        self._pubsub: int
+        self._server: Any
+
     @command((bytes,))
-    def echo(self, message):
+    def echo(self, message: bytes) -> bytes:
         return message
 
     @command((), (bytes,))
-    def ping(self, *args):
+    def ping(self, *args: bytes) -> Union[List[bytes], bytes, SimpleString]:
         if len(args) > 1:
             msg = msgs.WRONG_ARGS_MSG6.format("ping")
             raise SimpleError(msg)
@@ -31,7 +33,7 @@ class ConnectionCommandsMixin:
             return args[0] if args else PONG
 
     @command((DbIndex,))
-    def select(self, index):
+    def select(self, index: DbIndex) -> SimpleString:
         self._db = self._server.dbs[index]
-        self._db_num = index
+        self._db_num = index  # type: ignore
         return OK