+++ fakeredis/commands_mixins/hash_mixin.py
@@ -1,19 +1,31 @@
 import itertools
 import math
 import random
-from typing import Callable
+from typing import Callable, List, Tuple, Any, Optional
 
 from fakeredis import _msgs as msgs
-from fakeredis._commands import command, Key, Hash, Int, Float
-from fakeredis._helpers import SimpleError, OK, casematch
+from fakeredis._commands import command, Key, Hash, Int, Float, CommandItem
+from fakeredis._helpers import SimpleError, OK, casematch, SimpleString
 
 
 class HashCommandsMixin:
     _encodeint: Callable[[int, ], bytes]
     _encodefloat: Callable[[float, bool], bytes]
+    _scan: Callable[[CommandItem, int, bytes, bytes], Tuple[int, List[bytes]]]
+
+    def _hset(self, key: CommandItem, *args: bytes) -> int:
+        h = key.value
+        keys_count = len(h.keys())
+        h.update(
+            dict(zip(*[iter(args)] * 2))  # type: ignore
+        )  # https://stackoverflow.com/a/12739974/1056460
+        created = len(h.keys()) - keys_count
+
+        key.updated()
+        return created
 
     @command((Key(Hash), bytes), (bytes,))
-    def hdel(self, key, *fields):
+    def hdel(self, key: CommandItem, *fields: bytes) -> int:
         h = key.value
         rem = 0
         for field in fields:
@@ -24,20 +36,20 @@ class HashCommandsMixin:
         return rem
 
     @command((Key(Hash), bytes))
-    def hexists(self, key, field):
+    def hexists(self, key: CommandItem, field: bytes) -> int:
         return int(field in key.value)
 
     @command((Key(Hash), bytes))
-    def hget(self, key, field):
+    def hget(self, key: CommandItem, field: bytes) -> Any:
         return key.value.get(field)
 
     @command((Key(Hash),))
-    def hgetall(self, key):
+    def hgetall(self, key: CommandItem) -> List[bytes]:
         return list(itertools.chain(*key.value.items()))
 
     @command(fixed=(Key(Hash), bytes, bytes))
-    def hincrby(self, key, field, amount):
-        amount = Int.decode(amount)
+    def hincrby(self, key: CommandItem, field: bytes, amount_bytes: bytes) -> int:
+        amount = Int.decode(amount_bytes)
         field_value = Int.decode(
             key.value.get(field, b"0"), decode_error=msgs.INVALID_HASH_MSG
         )
@@ -47,7 +59,7 @@ class HashCommandsMixin:
         return c
 
     @command((Key(Hash), bytes, bytes))
-    def hincrbyfloat(self, key, field, amount):
+    def hincrbyfloat(self, key: CommandItem, field: bytes, amount: bytes) -> bytes:
         c = Float.decode(key.value.get(field, b"0")) + Float.decode(amount)
         if not math.isfinite(c):
             raise SimpleError(msgs.NONFINITE_MSG)
@@ -57,30 +69,24 @@ class HashCommandsMixin:
         return encoded
 
     @command((Key(Hash),))
-    def hkeys(self, key):
+    def hkeys(self, key: CommandItem) -> List[bytes]:
         return list(key.value.keys())
 
     @command((Key(Hash),))
-    def hlen(self, key):
+    def hlen(self, key: CommandItem) -> int:
         return len(key.value)
 
     @command((Key(Hash), bytes), (bytes,))
-    def hmget(self, key, *fields):
+    def hmget(self, key: CommandItem, *fields: bytes) -> List[bytes]:
         return [key.value.get(field) for field in fields]
 
     @command((Key(Hash), bytes, bytes), (bytes, bytes))
-    def hmset(self, key, *args):
+    def hmset(self, key: CommandItem, *args: bytes) -> SimpleString:
         self.hset(key, *args)
         return OK
 
-    @command(
-        (
-                Key(Hash),
-                Int,
-        ),
-        (bytes, bytes),
-    )
-    def hscan(self, key, cursor, *args):
+    @command((Key(Hash), Int), (bytes, bytes))
+    def hscan(self, key: CommandItem, cursor: int, *args: bytes) -> List[Any]:
         cursor, keys = self._scan(key.value, cursor, *args)
         items = []
         for k in keys:
@@ -89,33 +95,25 @@ class HashCommandsMixin:
         return [cursor, items]
 
     @command((Key(Hash), bytes, bytes), (bytes, bytes))
-    def hset(self, key, *args):
-        h = key.value
-        keys_count = len(h.keys())
-        h.update(
-            dict(zip(*[iter(args)] * 2))
-        )  # https://stackoverflow.com/a/12739974/1056460
-        created = len(h.keys()) - keys_count
-
-        key.updated()
-        return created
+    def hset(self, key: CommandItem, *args: bytes) -> int:
+        return self._hset(key, *args)
 
     @command((Key(Hash), bytes, bytes))
-    def hsetnx(self, key, field, value):
+    def hsetnx(self, key: CommandItem, field: bytes, value: bytes) -> int:
         if field in key.value:
             return 0
-        return self.hset(key, field, value)
+        return self._hset(key, field, value)
 
     @command((Key(Hash), bytes))
-    def hstrlen(self, key, field):
+    def hstrlen(self, key: CommandItem, field: bytes) -> int:
         return len(key.value.get(field, b""))
 
     @command((Key(Hash),))
-    def hvals(self, key):
+    def hvals(self, key: CommandItem) -> List[bytes]:
         return list(key.value.values())
 
     @command(name="HRANDFIELD", fixed=(Key(Hash),), repeat=(bytes,))
-    def hrandfield(self, key, *args):
+    def hrandfield(self, key: CommandItem, *args: bytes) -> Optional[List[bytes]]:
         if len(args) > 2:
             raise SimpleError(msgs.SYNTAX_ERROR_MSG)
         if key.value is None or len(key.value) == 0:
@@ -135,6 +133,3 @@ class HashCommandsMixin:
         else:
             res = [t[0] for t in res]
         return res
-
-    def _scan(self, keys, cursor, *args):
-        raise NotImplementedError  # Implemented in BaseFakeSocket