+++ fakeredis/commands_mixins/scripting_mixin.py
@@ -2,16 +2,18 @@ import functools
 import hashlib
 import itertools
 import logging
-from typing import Tuple, Callable, AnyStr, Set, Any
+from typing import Callable, AnyStr, Set, Any, Tuple, List, Dict
+
+from lupa import LuaRuntime
 
 from fakeredis import _msgs as msgs
-from fakeredis._commands import command, Int
+from fakeredis._commands import command, Int, Signature
 from fakeredis._helpers import (
     SimpleError,
     SimpleString,
     null_terminate,
     OK,
-    encode_command,
+    decode_command_bytes,
 )
 
 LOGGER = logging.getLogger("fakeredis")
@@ -29,15 +31,15 @@ REDIS_LOG_LEVELS_TO_LOGGING = {
 }
 
 
-def _ensure_str(s: AnyStr, encoding: str, replaceerr: str):
+def _ensure_str(s: AnyStr, encoding: str, replaceerr: str) -> str:
     if isinstance(s, bytes):
         res = s.decode(encoding=encoding, errors=replaceerr)
     else:
-        res = str(s).encode(encoding=encoding, errors=replaceerr)
+        res = str(s)
     return res
 
 
-def _check_for_lua_globals(lua_runtime, expected_globals):
+def _check_for_lua_globals(lua_runtime: LuaRuntime, expected_globals: Set[Any]) -> None:
     unexpected_globals = set(lua_runtime.globals().keys()) - expected_globals
     if len(unexpected_globals) > 0:
         unexpected = [
@@ -46,7 +48,7 @@ def _check_for_lua_globals(lua_runtime, expected_globals):
         raise SimpleError(msgs.GLOBAL_VARIABLE_MSG.format(", ".join(unexpected)))
 
 
-def _lua_redis_log(lua_runtime, expected_globals, lvl, *args):
+def _lua_redis_log(lua_runtime: LuaRuntime, expected_globals: Set[Any], lvl: int, *args: Any) -> None:
     _check_for_lua_globals(lua_runtime, expected_globals)
     if len(args) < 1:
         raise SimpleError(msgs.REQUIRES_MORE_ARGS_MSG.format("redis.log()", "two"))
@@ -63,16 +65,16 @@ def _lua_redis_log(lua_runtime, expected_globals, lvl, *args):
 
 
 class ScriptingCommandsMixin:
-    version: Tuple[int]
-    _name_to_func: Callable
-    _run_command: Callable
+    _name_to_func: Callable[[str, ], Tuple[Callable[..., Any], Signature]]
+    _run_command: Callable[[Callable[..., Any], Signature, List[Any], bool], Any]
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, *args: Any, **kwargs: Any):
         super(ScriptingCommandsMixin, self).__init__(*args, **kwargs)
         # Maps SHA1 to the script source
-        self.script_cache = {}
+        self.script_cache: Dict[bytes, bytes] = {}
+        self.version: Tuple[int]
 
-    def _convert_redis_arg(self, lua_runtime, value):
+    def _convert_redis_arg(self, lua_runtime: LuaRuntime, value: Any) -> bytes:
         # Type checks are exact to avoid issues like bool being a subclass of int.
         if type(value) is bytes:
             return value
@@ -87,7 +89,7 @@ class ScriptingCommandsMixin:
             )
             raise SimpleError(msg)
 
-    def _convert_redis_result(self, lua_runtime, result):
+    def _convert_redis_result(self, lua_runtime: LuaRuntime, result: Any) -> Any:
         if isinstance(result, (bytes, int)):
             return result
         elif isinstance(result, SimpleString):
@@ -108,7 +110,7 @@ class ScriptingCommandsMixin:
                 "Unexpected return type from redis: {}".format(type(result))
             )
 
-    def _convert_lua_result(self, result, nested=True):
+    def _convert_lua_result(self, result: Any, nested: bool = True) -> Any:
         from lupa import lua_type
 
         if lua_type(result) == "table":
@@ -139,22 +141,22 @@ class ScriptingCommandsMixin:
             return 1 if result else None
         return result
 
-    def _lua_redis_call(self, lua_runtime, expected_globals, op, *args):
+    def _lua_redis_call(self, lua_runtime: LuaRuntime, expected_globals: Set[Any], op: bytes, *args: Any) -> Any:
         # Check if we've set any global variables before making any change.
         _check_for_lua_globals(lua_runtime, expected_globals)
-        func, sig = self._name_to_func(encode_command(op))
+        func, sig = self._name_to_func(decode_command_bytes(op))
         new_args = [self._convert_redis_arg(lua_runtime, arg) for arg in args]
         result = self._run_command(func, sig, new_args, True)
         return self._convert_redis_result(lua_runtime, result)
 
-    def _lua_redis_pcall(self, lua_runtime, expected_globals, op, *args):
+    def _lua_redis_pcall(self, lua_runtime: LuaRuntime, expected_globals: Set[Any], op: bytes, *args: Any) -> Any:
         try:
             return self._lua_redis_call(lua_runtime, expected_globals, op, *args)
         except Exception as ex:
             return lua_runtime.table_from({b"err": str(ex)})
 
     @command((bytes, Int), (bytes,), flags=msgs.FLAG_NO_SCRIPT)
-    def eval(self, script, numkeys, *keys_and_args):
+    def eval(self, script: bytes, numkeys: int, *keys_and_args: bytes) -> Any:
         from lupa import LuaError, LuaRuntime, as_attrgetter
 
         if numkeys > len(keys_and_args):
@@ -163,7 +165,7 @@ class ScriptingCommandsMixin:
             raise SimpleError(msgs.NEGATIVE_KEYS_MSG)
         sha1 = hashlib.sha1(script).hexdigest().encode()
         self.script_cache[sha1] = script
-        lua_runtime = LuaRuntime(encoding=None, unpack_returned_tuples=True)
+        lua_runtime: LuaRuntime = LuaRuntime(encoding=None, unpack_returned_tuples=True)
 
         set_globals = lua_runtime.eval(
             """
@@ -206,21 +208,16 @@ class ScriptingCommandsMixin:
 
         return self._convert_lua_result(result, nested=False)
 
-    @command((bytes, Int), (bytes,), flags=msgs.FLAG_NO_SCRIPT)
-    def evalsha(self, sha1, numkeys, *keys_and_args):
+    @command(name="EVALSHA", fixed=(bytes, Int), repeat=(bytes,), flags=msgs.FLAG_NO_SCRIPT)
+    def evalsha(self, sha1: bytes, numkeys: Int, *keys_and_args: bytes) -> Any:
         try:
             script = self.script_cache[sha1]
         except KeyError:
             raise SimpleError(msgs.NO_MATCHING_SCRIPT_MSG)
         return self.eval(script, numkeys, *keys_and_args)
 
-    @command(
-        name="script load",
-        fixed=(bytes,),
-        repeat=(bytes,),
-        flags=msgs.FLAG_NO_SCRIPT,
-    )
-    def script_load(self, *args):
+    @command(name="SCRIPT LOAD", fixed=(bytes,), repeat=(bytes,), flags=msgs.FLAG_NO_SCRIPT)
+    def script_load(self, *args: bytes) -> bytes:
         if len(args) != 1:
             raise SimpleError(msgs.BAD_SUBCOMMAND_MSG.format("SCRIPT"))
         script = args[0]
@@ -228,24 +225,14 @@ class ScriptingCommandsMixin:
         self.script_cache[sha1] = script
         return sha1
 
-    @command(
-        name="script exists",
-        fixed=(),
-        repeat=(bytes,),
-        flags=msgs.FLAG_NO_SCRIPT,
-    )
-    def script_exists(self, *args):
+    @command(name="SCRIPT EXISTS", fixed=(), repeat=(bytes,), flags=msgs.FLAG_NO_SCRIPT)
+    def script_exists(self, *args: bytes) -> List[int]:
         if self.version >= (7,) and len(args) == 0:
             raise SimpleError(msgs.WRONG_ARGS_MSG7)
         return [int(sha1 in self.script_cache) for sha1 in args]
 
-    @command(
-        name="script flush",
-        fixed=(),
-        repeat=(bytes,),
-        flags=msgs.FLAG_NO_SCRIPT,
-    )
-    def script_flush(self, *args):
+    @command(name="SCRIPT FLUSH", fixed=(), repeat=(bytes,), flags=msgs.FLAG_NO_SCRIPT)
+    def script_flush(self, *args: bytes) -> SimpleString:
         if len(args) > 1 or (
                 len(args) == 1 and null_terminate(args[0]) not in {b"sync", b"async"}
         ):
@@ -254,11 +241,11 @@ class ScriptingCommandsMixin:
         return OK
 
     @command((), flags=msgs.FLAG_NO_SCRIPT)
-    def script(self, *args):
+    def script(self, *args: bytes) -> None:
         raise SimpleError(msgs.BAD_SUBCOMMAND_MSG.format("SCRIPT"))
 
     @command(name="SCRIPT HELP", fixed=())
-    def script_help(self, *args):
+    def script_help(self, *args: bytes) -> List[bytes]:
         help_strings = [
             "SCRIPT <subcommand> [<arg> [value] [opt] ...]. Subcommands are:",
             "DEBUG (YES|SYNC|NO)",